pyproject.toml
[project]
name = "grpc-service"
version = "0.1.0"
description = "gRPC service with Protocol Buffers"
requires-python = ">=3.9"
dependencies = [
    "grpcio>=1.68.1",
    "grpcio-tools>=1.68.1",
    "mypy-protobuf>=3.6.0",
    "mypy>=1.13.0",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["grpc.*"]
ignore_missing_imports = true

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "B", "I"]
